requirements.txt
flask
vertexai
google-generativeai